backend/pyproject.toml
[project]
name = "ai-docvivid-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "celery>=5.5.3",
    "fastapi[standard]>=0.116.1",
    "google-cloud-storage>=3.4.1",
    "google-genai>=1.44.0",
    "greenlet>=3.2.3",
    "python-jose[cryptography]==3.3.0",
    "jose==1.0.0",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.28",
    "moviepy>=2.2.1",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "tenacity>=9.1.2",
    "tiktoken>=0.12.0",
    "beautifulsoup4>=4.12.3",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<5.0.0",
    "markitdown[all]==0.1.3",
    "bagelpay>=1.0.3",
]
